+++ src/components/BloodChart.js
@@ -4,8 +4,7 @@ import Loading from './Loading';
 
 
 import { getOnePatient } from '@/data/getData';
-import PatientChart from './Chart';
-
+import PatientChart from "./Chart/Chart";
 
 const BloodChart = async () => {
