--- src/components/DiagnosticList.js
@@ -1,11 +1,0 @@
-
-
-import CardTitle from "@/components/CardTitle";
-
-const DiagnosticList = () => (
-    <div class="flex flex-col bg-white rounded-xl ">
-        <CardTitle  title='Diagnostic List'></CardTitle>
-    </div>
-);
-
-export default DiagnosticList
